Move files inside collison into specific folder and fix cmake path
