+++ src/common/Collision/Management/VMapFactory.cpp
