+++ src/common/Collision/Management/VMapFactory.h
