+++ src/common/Collision/Models/GameObjectModel.cpp
@@ -15,8 +15,8 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "VMapFactory.h"
-#include "VMapManager2.h"
+#include "Management/VMapFactory.h"
+#include "Management/VMapManager2.h"
 #include "VMapDefinitions.h"
 #include "WorldModel.h"
