+++ src/common/Collision/Models/GameObjectModel.h
