+++ src/common/Collision/Models/ModelInstance.cpp
@@ -17,7 +17,7 @@
 
 #include "ModelInstance.h"
 #include "WorldModel.h"
-#include "MapTree.h"
+#include "Maps/MapTree.h"
 #include "VMapDefinitions.h"
 
 using G3D::Vector3;
