+++ src/server/game/Map.cpp
@@ -19,7 +19,7 @@
 #include "MapManager.h"
 #include "GridStates.h"
 #include "ScriptMgr.h"
-#include "VMapFactory.h"
+#include "Management/VMapFactory.h"
 #include "MapInstanced.h"
 #include "CellImpl.h"
 #include "GridNotifiers.h"
