+++ src/tools/mmaps_generator/TerrainBuilder.h
@@ -22,7 +22,7 @@
 #include "MangosMap.h"
 #include "MoveMapSharedDefines.h"
 
-#include "WorldModel.h"
+#include "Models/WorldModel.h"
 
 #include "G3D/Array.h"
 #include "G3D/Vector3.h"
