+++ src/tools/vmap4_assembler/VMapAssembler.cpp
@@ -18,7 +18,7 @@
 #include <string>
 #include <iostream>
 
-#include "TileAssembler.h"
+#include "Maps/TileAssembler.h"
 
 int main(int argc, char* argv[])
 {
